+++ plugins/com.google.cloud.tools.eclipse.appengine.facets.test/src/com/google/cloud/tools/eclipse/appengine/facets/AppEngineFlexFacetTest.java
@@ -1,4 +1,4 @@
-package com.google.cloud.tools.eclipse.appengine.facets.flex;
+package com.google.cloud.tools.eclipse.appengine.facets;
 
 import org.eclipse.wst.common.project.facet.core.ProjectFacetsManager;
 import org.junit.Assert;
